requirements.txt
# Dev environment
pip
autopep8

# App
flask
jsonify
requests
flask-cors
python-dotenv